Youtube Comment Sentiment Analysis Final Project
The final Project on YouTube comment sentiment analysis project file was added to the repo. The files consist of Project Proposal, Report, Presentation slides, ipynb notebook file, and its pdf file. It also has a comprehensive README file that describes the project that was done as a part of the course.